redone ui

+++ server/profile_service/src/main/java/com/example/profile_service/repository/UserRepo.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepo extends JpaRepository<User, Long>  {
+public interface UserRepo extends JpaRepository<User, Integer>  {
 
     User            findByEmail(String email);
     List<User>      findByRole(Role role);
